- Serves a valid xml rss podcast feed for Skumma kultur w dummy data.
But under the hood it's actually hooked up to the digas db and is capable
of fetching podcasts episodes straigth from digas.

Also we have a start on implementing programinfo. I plan to have
a import method from the old server. Dumped json files from the old db.
import json.

- notice there are design decitions to be made about url-encoding data.
  - in the db?
  - post/pre db?